requirements.txt
streamlit
langchain
openai
langchain-community
langchain-openai
python-dotenv